// oms-ui-ng2/src/app/common/models/user.ts
import { Role } from './role';
export class User {
  firstName: string;
  lastName: string;
  username: string;
  emailId: string;
  address1: string;
  address2: string;
  state: string;
  city: string;
  zip: string;
  country: string;
  telelphoneNo: string;
  mobileNo: string;
  faxNo: string;
  enabled: boolean;
  role: Role;
  active:boolean;
}